requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.7.0rc2
dj-rest-auth==2.1.3
Django==3.1.5
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-phonenumber-field==5.0.0
django-rest-auth==0.9.5
django-richtextfield==1.6
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack==1.0.2
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.0.1
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
uritemplate==3.0.1
urllib3==1.26.2
zope.interface==5.2.0
